second commit

+++ src/main/java/com/example/ecomerce/website/models/Images.java
@@ -19,11 +19,11 @@ public class Images {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	@Column(name="id")
-	private int id;
-	public int getId() {
+	private long id;
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	@ManyToOne(fetch=FetchType.EAGER,cascade={CascadeType.DETACH,
